+++ module_prototyper/models/__init__.py
@@ -2,7 +2,7 @@
 ##############################################################################
 #
 #    OpenERP, Open Source Management Solution
-#    This module copyright (C) 2010 - 2014 Savoir-faire Linux
+#    This module copyright (C) 2010 - 2014, 2017 Savoir-faire Linux
 #    (<http://www.savoirfairelinux.com>).
 #
 #    This program is free software: you can redistribute it and/or modify
@@ -22,5 +22,4 @@
 from . import (
     module_prototyper,
     ir_model_fields,
-    topo
 )
