+++ module_prototyper/models/module_prototyper.py
@@ -35,7 +35,7 @@ from openerp import models, api, fields
 from openerp.tools.safe_eval import safe_eval
 
 from . import licenses
-from .topo import find_depends, MAGIC_COLUMNS
+from ..utils.closure import find_closure, NON_CLOSURE_COLUMNS as MAGIC_COLUMNS
 
 _logger = logging.getLogger(__name__)
 
@@ -452,7 +452,7 @@ class ModulePrototyper(models.Model):
                 fname = self.friendly_name(self.unprefix(model_name))
                 filename = '{0}/{1}.xml'.format(prefix, fname)
                 self._data_files.append(filename)
-                dependencies = find_depends(records)
+                dependencies = find_closure(records)
                 res.append(self.generate_file_details(
                     filename,
                     'data/model_name.xml.template',
